Add exception handling to unit tests
