+++ build.gradle
@@ -166,6 +166,7 @@ dependencies {
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter', version: versions.junitJupiter
     testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test'
 
+    testImplementation 'org.mockito:mockito-junit-jupiter:2.23.0'
     testImplementation 'ru.lanwen.wiremock:wiremock-junit5:1.3.0'
 
     testImplementation group: 'com.github.tomakehurst', name: 'wiremock', version: '2.24.1'
