+++ src/test/java/uk/gov/hmcts/reform/payments/client/PaymentsClientTest.java
@@ -3,129 +3,75 @@ package uk.gov.hmcts.reform.payments.client;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
-import org.springframework.stereotype.Service;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.web.bind.annotation.RequestMethod;
-import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.payments.client.models.FeeDto;
-import uk.gov.hmcts.reform.payments.client.models.PaymentDto;
 
 import java.math.BigDecimal;
-import java.util.Map;
 
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
-@SpringBootTest(classes = {
-        PaymentsApi.class,
-        PaymentsClient.class,
-        PaymentsClientTest.ServiceTestSupportAuthTokenGenerator.class,
-        PaymentsClientTest.StubServiceAuthorisationImpl.class
-})
-@ExtendWith(SpringExtension.class)
-@EnableAutoConfiguration
-@AutoConfigureWireMock(port = 8091)
+@ExtendWith(MockitoExtension.class)
 class PaymentsClientTest {
-    @Autowired
-    private ServiceTestSupportAuthTokenGenerator authTokenGenerator;
+    private static final CardPaymentRequest PAYMENT_REQUEST = CardPaymentRequest.builder()
+            .amount(BigDecimal.valueOf(999.99))
+            .caseReference("case reference")
+            .ccdCaseNumber("ccd case number")
+            .channel("channel")
+            .currency("currency")
+            .description("description")
+            .provider("provider")
+            .service("service")
+            .siteId("site ID")
+            .build();
 
-    @Autowired
+    @Mock
     private PaymentsApi paymentsApi;
 
-    private PaymentsClient paymentsClient;
+    @Mock
+    private AuthTokenGenerator authTokenGenerator;
+
+    private PaymentsClient client;
 
     @BeforeEach
-    void setupClient() {
-        this.paymentsClient = new PaymentsClient(paymentsApi, authTokenGenerator);
+    void setUp() {
+        client = new PaymentsClient(paymentsApi, authTokenGenerator);
+        when(authTokenGenerator.generate()).thenReturn("auth token");
     }
 
     @Test
-    void testCreatePayment() {
-        PaymentDto payment = paymentsClient.createPayment(
-                "Authorisation",
-                CardPaymentRequest.builder()
-                        .amount(BigDecimal.TEN)
-                        .caseReference("e1880a05-740e-4a71-a82a-4931cc09f790")
-                        .ccdCaseNumber("UNKNOWN")
-                        .currency("GBP")
-                        .description("description")
-                        .fees(new FeeDto[0])
-                        .service("my service")
-                        .siteId("AA00")
-                        .build(),
-                "http://localhost"
-        );
-        assertNotNull(payment);
-        assertAll(
-            () -> assertEquals("RC-1566-2080-1331-6610", payment.getReference()),
-            () -> assertEquals("Initiated", payment.getStatus()),
-            () -> assertEquals("3pev48jhsin6n1sr91lgvi9j3o", payment.getExternalReference()),
-            () -> assertEquals("2019-15662080132", payment.getPaymentGroupReference()),
-            () -> assertNotNull(payment.getLinks()),
-            () -> assertAll(
-                () -> assertNotNull(payment.getLinks().getNextUrl()),
-                () -> assertEquals(
-                        "https://www.payments.service.gov.uk/secure/473320d7-4db4-42a5-935f-8bcaf0bd2056",
-                        payment.getLinks().getNextUrl().getHref().toString()
-                ),
-                () -> assertEquals(RequestMethod.GET, payment.getLinks().getNextUrl().getMethod())
-            )
-        );
+    void createPaymentShouldInvokePaymentsApi() {
+        client.createPayment("authorisation", PAYMENT_REQUEST, "redirect");
+
+        verify(paymentsApi).create("authorisation", "auth token", "redirect", PAYMENT_REQUEST);
     }
 
     @Test
-    void testRetrievePayment() {
-        PaymentDto payment = paymentsClient.retrievePayment("Authorisation", "RC-1566-2093-5462-0545");
-        assertNotNull(payment);
-        assertAll(
-            () -> assertEquals("RC-1566-2093-5462-0545", payment.getReference()),
-            () -> assertEquals(BigDecimal.TEN, payment.getAmount()),
-            () -> assertEquals("description", payment.getDescription()),
-            () -> assertEquals("GBP", payment.getCurrency()),
-            () -> assertEquals("online", payment.getChannel()),
-            () -> assertEquals("card", payment.getMethod()),
-            () -> assertEquals("gov pay", payment.getExternalProvider()),
-            () -> assertNotNull(payment.getLinks()),
-            () -> assertAll(
-                () -> assertEquals(
-                        "http://localhost:4421/card-payments/RC-1566-2093-5462-0545",
-                        payment.getLinks().getSelf().getHref().toString()
-                ),
-                () -> assertEquals(RequestMethod.GET, payment.getLinks().getSelf().getMethod())
-            )
-        );
+    void retrievePaymentShouldInvokePaymentsApi() {
+        client.retrievePayment("authorisation", "payment reference");
+
+        verify(paymentsApi).retrieve("payment reference", "authorisation", "auth token");
     }
 
-    @Service
-    static class ServiceTestSupportAuthTokenGenerator implements AuthTokenGenerator {
-        @Override
-        public String generate() {
-            return "let me in";
-        }
+    @Test
+    void createPaymentShouldPropagateExceptions() {
+        when(authTokenGenerator.generate())
+                .thenThrow(new RuntimeException("expected exception for create payment"));
+
+        assertThatThrownBy(() -> client.createPayment("authorisation", PAYMENT_REQUEST, "redirect"))
+                .isInstanceOf(RuntimeException.class)
+                .hasMessage("expected exception for create payment");
     }
 
-    @Service
-    static class StubServiceAuthorisationImpl implements ServiceAuthorisationApi {
+    @Test
+    void retrievePaymentShouldPropagateExceptions() {
+        when(authTokenGenerator.generate())
+                .thenThrow(new RuntimeException("expected exception for retrieve payment"));
 
-        @Override
-        public String serviceToken(Map<String, String> signIn) {
-            return null;
-        }
-
-        @Override
-        public void authorise(String authHeader, String[] roles) {
-
-        }
-
-        @Override
-        public String getServiceName(String authHeader) {
-            return null;
-        }
+        assertThatThrownBy(() -> client.retrievePayment("authorisation", "payment reference"))
+                .isInstanceOf(RuntimeException.class)
+                .hasMessage("expected exception for retrieve payment");
     }
 }
